[avalanche] Rationalize the local proof status in getavalancheinfo
Summary:
The `getavalancheinfo` RPC returns several info about the local proof, including a `verified` flag, a `verification_status` and a `sharing` flag.
This information has proven to be confusing, as the `sharing` flag doesn't give any information about why the proof is not shared, and one of the verification status is `pending` which doesn't indicate why the verification is pending.

This diffs deprecates the `sharing` flag in favor of a better verification status, so pending can be either `pending verification`  or `pending inbound connections` which is equivalent to the previous `sharing: false` field.

Depends on D14565.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14566