Lighten the net logging
Summary:
Most of the logs from the net category are either errors/unusual behavior messages or important network event like version message, new block download etc.
The exception is the couple "received: " and "sending: " messages that log all the traffic in and out. It was OK until avalanche, and now it causes the log to be bloated by avapoll and avaresponse messages.

This diff creates a new log category "netdebug" that only contains these 2 messages for now, leaving the net logging almost identical but less bloated.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, bytesofman, PiRK

Reviewed By: #bitcoin_abc, bytesofman, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14573