[Cashtab] Refactor new tx function to accept targetOutputs
Summary:
Have new `sendXec` function accept targetOutputs instead of creating it. This will allow extended support of different types of XEC txs without overloading this sendXec function.

While the function is renamed to be more general, it is not yet unit tested for other cases. These will be added before the function is implemented in the app.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14574