[Cashtab] Remove encrypted msg sending
Summary:
Remove feature for sending encrypted msgs. Deprecation cycle has been short, but we have only seen a few of these msgs since August, last one on Oct 3.

Initial idea here was to show it can be done when the main focus of Cashtab development was trying out new features and iterating toward unique web wallet use cases. The feature is high complexity and not widely used. Supporting it through other Cashtab upgrades is not worth it.

Ultimately, if you want a message to be secret -- the blockchain is not the place for it.

This diff removes the ability to send encrypted msg op_return txs in Cashtab. However, functions used to parse such msgs in tx history, and functions used to create the actual transactions, are left unchanged for now. This is because the functions that deal with encrypted msg sending are already-overloaded tx creation functions that will be replaced by new functions. It's not worth removing this complexity from these functions as they will be rewritten and dropped in other diffs.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14577