[electrum] Fix various strings that were wrong, inconsistent or not translatable
Summary:
The titles of load / save dialogs used the wrong title in various
locations, this fixes it.

This also makes some previously untranlatable string translatable.

This is a backport of [[https://github.com/Electron-Cash/Electron-Cash/pull/2718 | electroncash#2718]]

Test Plan:
In the contacts tab, right-click  and select "Import contacts"

I'm not sure how to reach the invoice importing dialog, but the fix is trivially correct.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14702