Replace ... single char with 3 dots
Summary:
This isn't an ascii char and it doesn't print good on all the terminals. The replacement doesn't even increase the binary size due to the encoding.
It is maintained as is in the RPC console though as it displays a bit nicer, and this is a controlled environment.

Test Plan:
  ninja all check-all
Grep for the char and see it has been removed everywhere it needs to be.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14633