[avalanche] Remember the dangling proofs
Summary:
Store them in a pool rather than a simpler filter on ids. This makes it possible to access the details of the dangling proofs.
The pool is limited in size the same way the other pools are, by the intrinsic utxo limitation. The only change of behavior is the refined log message.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK, bytesofman

Reviewed By: #bitcoin_abc, PiRK, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14658