util: move HexStr and HexDigit from util to crypto
Summary:
Move HexStr and HexDigit functions from util to crypto. The crypto library does
not actually use these functions, but the consensus library does. The consensus
and util libraries not allowed to depend on each other, but are allowed to
depend on the cryto library, so the crypto library is a reasonable put these.

The consensus library uses HexStr and HexDigit in script.cpp, transaction.cpp,
and uint256.cpp.

The util library does not use HexStr but does use HexDigit in strencodings.cpp
to parse integers.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29015 | core#29015]]
https://github.com/bitcoin/bitcoin/pull/29015/changes/23cc8ddff472d259605d7790ba98a1900e77efab
Depends on D19743

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19744