[Prettier] Use prettier on Cashtab extension files
Summary:
Remove 'extension' folder from .prettierignore

Prettier should be used to lint the js files in this repo. This was originally added so that the `dist/` folder in `extension/` would not get the prettier treatment. However, this is already covered by `dist` being included in .prettierignore.

Test Plan: `arc lint`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14693