[e.cash] getScores edit
Summary:
The min-max score checker for exchanges was resolving null values as 0 and therefore meeting the threshold as true if the
min check was 0. This was causing a handful of scores to be higher than they should be.
Added a check to ignore null values so the scores are computed properly, and updated the test data to reflect this

Test Plan:
Preview the app and check the get-ecash page renders properly and the order of the items matches scorecard.cash, specifically the instant exchanges
check tests still pass - npm run test

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14705