[Cashtab][Alias] Check Alias function
Summary:
Simplified and more efficient version of D14666 where a `Check Alias` button is added which leverages the existing registration input field and validation logic.
Results from the check are rendered via UI notifications.

Clicking Check Alias on an alias with pending registrations
{F7088795}

Clicking Check Alias on an alias with no pending registrations
{F7088797}

When an invalid alias is entered, the Check Alias button is also disabled
{F7088798}

Test Plan:
npm test
npm start
Check on a registered, unregistered and pending alias and observe the UI notifications on their status
Input an invalid alias and ensure the Check Alias button is disabled

Reviewers: bytesofman, Fabien, #bitcoin_abc

Reviewed By: bytesofman, Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D14713