[avalanche] Fix the return status of the setstakingreward RPC
Summary: The expected behavior is to return true if the payout script was set successfully. However the actual return value comes from the std::unordered_map::insert_or_assign that returns true on insertion but false on assignment. This diff fixes the return value to be compliant with the expected return value for the RPC.

Test Plan:
  ./test/functional/test_runner.py abc_mining_stakingrewards

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14714