[avalanche] Consolidate the proofs recent finalization state memorization
Summary: D14704 fixed a bug regarding the recently finalized proofs status that could be lost under some circumstances. This diff is a follow up that applies the same logic consistently to all the places in our codebase. Note that the places where it was not used are places where it was not needed, but it is cheap enough and it guarantees that we will not be missing some edge cases in the future.

Test Plan:
With Clang and Debug:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14732