[CI] Fix the Win64 build failing due to wine issue
Summary:
Wine has some limitations, and as the test suite grows so does the memory requirement. With the current Bullseye version we are facing a memory limit causing asserts on nullptr for objects that failed to allocate.
We can work around this limitation by using a more recent version of wine. This also speeds up the tests significantly on my machine.

Test Plan:
Install the correct wine version, then:
  ./contrib/teamcity/build-configurations.py build-win64

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14751