[CI] Unbreak the win64 build
Summary:
The buildis flaky due to an assertion occuring about 50% of the time in the denial of service test. It can only be reproduced when all the tests are run, not on a single test or even a subset.
A proper fix is being investigated, but for now disable the failing test so we don't introduce new issues silently.

Test Plan:
  ./contrib/teamcity/build-configurations.py build-win64

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14770