[CI] Double the functional tests timeout for coverage and emulator builds
Summary: This is an attempt to make the functional tests more resilient on very slow builds by doubling the timeout specifically for the builds that require it. This is intended to make the CI less flaky for these builds.

Test Plan:
  ninja check-functional
Check the timeout is doubled with quemu and coverage, not regular builds.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14782