[CI] Unbreak the wine installation
Summary:
Apt cannot resolve the dependencies of a version specific package if it's not the latest version, so we need to pinpoint all the dependencies as well.
This will unbreak building the container, which is not yet an issue on CI due to apt caching but prevents from building the docker dev image.

Test Plan:
Build the docker dev image.
Check the CI runs on this diff.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14797