[chronik-client] Generate proto file for in-node chronik
Summary:
The proto file for the in-node version of chronik is generated from the in-node proto file in the monorepo.

Note: The legacy `build-proto` script is dependent on running the command from a local repo with `bitcoinsuite` in the correct place. It should not remain in the monorepo codebase. Updates to the NNG chronik version are not anticipated in the monorepo. If they do occur, the dependencies need to be pulled in in some other way anyway, so the script should still go.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14814