[Cashtab] Do not try to parse outputscript of coinbase txs
Summary:
Patch bug in Cashtab tx history parsing causing an error to be thrown in wallets with coinbase txs

Note: there is a lot more we can do for these txs, i.e. rendering them as such. This diff is minimal as the issue currently causes Cashtab wallets with any coinbase tx in the last 10 txs to be unusable.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14818