[Cashtab] app.js format improvements
Summary:
Match extension and non-extension app.js files where they are not expected to diverge

Have been evaluating removing the use of two separate files for extension and non-extension, and writing the code so that the files are parsed differently at buildtime e.g. using an env variable to build the extension. Running into some barriers with this approach but, either way, there is some whitespace to remove in the extension file, and legacy capitalization to fix in the non-extension file

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14824