[electrum] don't show error if the Aux Key Dialog's  password prompt is cancelled
Summary:
If the user opened the Aux Key dialog via the menu or the "Genereate Key" button in the Delegation Editor, and then cancelled the password prompt, it showed an error popup with a python stack trace. Handle this situation and just show a warning with an informative message.

Also don't set the suggested_delegated_key_and_index attribute in the Delegation editor, to avoid incrementing the aux key index if the user decides to later paste a  public key from an unknown source and generate a delegation.

Test Plan: Show the aux key dialog (both via the wallet menu and the delegation editor) in an encrypted wallet and cancel the password prompt.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14842