[seeder] handle CAddress-without-nTime logic
Summary: This is a follow-up bugfix to D14753. The CAddress serialization code was changed in the node to no longer handle the special case of CAddress-without-nTime that is specific to the version message, but the seeder was not updated accordingly. This should fix it.

Test Plan: Run the seeder, see if it finds valid peers.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14856